Wrap 'Failure' by 'TestResult' enum.
