+++ src/core/mod.rs
@@ -4,12 +4,12 @@
 pub use self::location::SourceLocation;
 pub use self::actual::{expect, ActualValue};
 pub use self::join::Join;
-pub use self::failure::{Failure, panics_with_message};
+pub use self::result::{TestResult, panics_with_message};
 
 mod location;
 mod actual;
 mod join;
-mod failure;
+mod result;
 
 /// A Matcher trait.
 pub trait Matcher<A, E> {
